Widgets refactored
