+++ lib/helper/common.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+bool isLandscape(BuildContext context) =>
+    MediaQuery.of(context).orientation == Orientation.landscape;
